+++ test/Props/Module2/Trait2.php
@@ -4,6 +4,8 @@ namespace Test\Exteon\Loader\ChainingClassResolver\Props\Module2;
 
 trait Trait2
 {
+    use Trait3;
+
     public function whoami(): string
     {
         return parent::whoami() . '+Trait2';