+++ test/Props/Module2/Trait3.php
@@ -1,0 +1,8 @@
+<?php
+
+    namespace Test\Exteon\Loader\ChainingClassResolver\Props\Module2;
+
+    trait Trait3
+    {
+
+    }