+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.1.0
+
+#### New features
+
+* Added `ChainedClassMeta::hasChainTrait()`
+
+#### Bugfixes
+
+* `ChainedClassMeta::getChainTraits()` was not traversing chained ancestor 
+  traits when it was called on a trait
+
 ### 1.0.2
 
 * Upgraded `exteon/mapping-class-loader`